Move healthcheck service
